requirements.txt
# Some example python packages
ephem
haversine
numpy
pillow
pytz
requests
